% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juno_palette.R
\name{juno_palette}
\alias{juno_palette}
\title{Juno colour palette}
\usage{
juno_palette()
}
\value{
A vector of hex colour codes
}
\description{
This function returns a custom colour palette for plots.
}
\examples{
\dontrun{
data <- data.frame(
x = rep(1:5, each = 3),
y = rnorm(15),
group = rep(letters[1:5], times = 3)
)
library(tidyverse)
ggplot(data, aes(x = x, y = y, colour = group)) +
geom_point(size = 4) +
scale_color_manual(values = juno_palette()) +
theme_minimal()
}
}
