% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_rayyan.R
\name{parse_rayyan}
\alias{parse_rayyan}
\title{parse_rayyan}
\usage{
parse_rayyan(rayyan_df)
}
\arguments{
\item{rayyan_df}{A rayyan exported bibliography file}
}
\value{
A dataframe
}
\description{
A function to parse bibliographic dataframes outputted from Rayyan, splitting up the 'notes' column into Inclusion/Exclusion decision, Labels, and Exclusion Reasons
}
\details{
function (beta) to remove any old inclusion labels from abstract screening (found in the labels or parsed as html paragraphs)
}
\references{
Jones ML & Grainger MJ (2021). rayyanR: An R package to process outputs of the Rayyan screening platform for systematic reviews. https://github.com/befriendabacterium/rayyanR.
}
\author{
Matt Lloyd Jones
}
